Revert "debugger: rename internal library for clarity"
This reverts commit d0c7ab00f9fcbf6ba45c3f554fa75cf60ca1b2ce.
