+++ lib/internal/inspector/_inspect.js
@@ -31,8 +31,8 @@ const util = require('util');
 
 const { 0: InspectClient, 1: createRepl } =
     [
-      require('internal/debugger/inspect_client'),
-      require('internal/debugger/inspect_repl'),
+      require('internal/inspector/inspect_client'),
+      require('internal/inspector/inspect_repl'),
     ];
 
 const debuglog = util.debuglog('inspect');
