+++ lib/internal/main/inspect.js
@@ -18,5 +18,5 @@ markBootstrapComplete();
 
 // Start the debugger agent.
 process.nextTick(() => {
-  require('internal/debugger/_inspect').start();
+  require('internal/inspector/_inspect').start();
 });
